refactor: 객체를 인스턴스화하는 역할을 담당하는 클래스를 분리한다.
