--- app/src/main/java/woowacourse/shopping/di/Instance.kt
@@ -1,9 +1,0 @@
-package woowacourse.shopping.di
-
-import kotlin.reflect.KClass
-import kotlin.reflect.full.superclasses
-
-data class Instance<T : Any>(val value: T) {
-
-    val clazz: List<KClass<*>> = value::class.superclasses + value::class
-}
