--- woogi-di/src/main/java/com/boogiwoogi/di/WoogiContainer.kt
@@ -1,10 +1,0 @@
-package com.boogiwoogi.di
-
-import kotlin.reflect.KClass
-
-interface WoogiContainer {
-
-    fun <T : Any> declareDependency(dependency: Dependency<T>)
-
-    fun find(clazz: KClass<*>): Any?
-}
